+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# NEXT
+- `KeyboardConfig` improvements. Now `directionalKeys` expect list of `KeyboardDirectionalKeys`
+
 # 3.8.3
 - Fix bug in the `PlatformPlayer` movements.
 - Adds `objectsBuilder` in `WorldMapBySpritefusion`. You can select a layer to adds objects in the tile position.